requirements.txt
tornado
sockjs-tornado
celery
tornado-celery